feat: move browser_*.js files to /lib
