+++ bin/template/cordova/Api.js
@@ -72,7 +72,7 @@ function Api (platform, platformRootDir, events) {
     if (!dirExists(this.root) || !dirExists(path.join(this.root, 'cordova'))) {
         throw new CordovaError('The provided path "' + this.root + '" is not a valid browser project.');
     }
-    this._handler = require('./browser_handler');
+    this._handler = require('../../../lib/browser_handler');
 
     this.locations = {
         platformRootDir,
