+++ lib/browser_handler.js
