+++ lib/browser_parser.js
